+++ ethereum/core/src/main/java/org/hyperledger/besu/ethereum/mainnet/MainnetProtocolSpecs.java
@@ -620,7 +620,7 @@ public abstract class MainnetProtocolSpecs {
       final boolean quorumCompatibilityMode,
       final EvmConfiguration evmConfiguration) {
 
-    return arrowGlacierDefinition(
+    return grayGlacierDefinition(
             chainId,
             configContractSizeLimit,
             configStackSizeLimit,