+++ ethereum/eth/src/main/java/org/hyperledger/besu/ethereum/eth/sync/backwardsync/SyncStepStep.java
@@ -50,7 +50,6 @@ public class SyncStepStep {
   private CompletableFuture<Block> requestBlock(final Hash targetHash) {
     final RetryingGetBlockFromPeersTask getBlockTask =
         RetryingGetBlockFromPeersTask.create(
-            context.getProtocolContext(),
             context.getProtocolSchedule(),
             context.getEthContext(),
             context.getMetricsSystem(),